Move Navigation.tsx under /Navigation sub-path
