+++ src/Pages/Authenticated/AuthenticatedPage.tsx
@@ -14,7 +14,7 @@ import Dashboard from './Dashboard/Dashboard';
 import Sessions from './Sessions/Sessions';
 import Settings from './Settings/Settings';
 import Wallet from './Wallet/Wallet';
-import Navigation from './Components/Navigation';
+import Navigation from './Navigation/Navigation';
 import { CircularProgress } from '@material-ui/core';
 import ArrowBackIcon from '@material-ui/icons/ArrowBack';
 import { Identity } from 'mysterium-vpn-js';
